Moved all refactored components into themed directory
