+++ src/components/scenes/WalletListScene.js
@@ -13,21 +13,21 @@ import XPubModal from '../../connectors/XPubModalConnector.js'
 import * as Constants from '../../constants/indexConstants.js'
 import s from '../../locales/strings.js'
 import { getDefaultIsoFiat, getIsAccountBalanceVisible } from '../../modules/Settings/selectors.js'
-import { WiredProgressBar } from '../../modules/UI/components/WiredProgressBar/WiredProgressBar.ui.js'
 import { getActiveWalletIds, getWalletLoadingPercent } from '../../modules/UI/selectors.js'
 import { type Dispatch, type RootState } from '../../types/reduxTypes.js'
 import { type GuiWallet } from '../../types/types.js'
 import { getTotalFiatAmountFromExchangeRates } from '../../util/utils.js'
 import { CrossFade } from '../common/CrossFade.js'
 import { SceneWrapper } from '../common/SceneWrapper.js'
-import { WalletList } from '../common/WalletList.js'
-import { WalletListFooter } from '../common/WalletListFooter.js'
-import { WalletListSortableRow } from '../common/WalletListSortableRow.js'
-import { WiredBalanceBox } from '../common/WiredBalanceBox.js'
 import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../services/ThemeContext.js'
 import { EdgeText } from '../themed/EdgeText.js'
 import { PromoCard } from '../themed/PromoCard.js'
 import { SettingsHeaderRow } from '../themed/SettingsHeaderRow.js'
+import { WalletList } from '../themed/WalletList.js'
+import { WalletListFooter } from '../themed/WalletListFooter.js'
+import { WalletListSortableRow } from '../themed/WalletListSortableRow.js'
+import { WiredBalanceBox } from '../themed/WiredBalanceBox.js'
+import { WiredProgressBar } from '../themed/WiredProgressBar.js'
 
 type StateProps = {
   activeWalletIds: string[],
